a few code cleanups

+++ match.c
@@ -56,7 +56,7 @@ static tag *tag_table;
 
 static int compare_targets(struct target *t1,struct target *t2)
 {
-  return(t1->t - t2->t);
+  return((int)t1->t - (int)t2->t);
 }
